Feat: Move socket code to composable
