+++ stores/locos.ts
@@ -1,6 +1,5 @@
 // import { Loco } from '@trainlink-org/shared-lib';
 import { Direction } from '@trainlink-org/shared-lib';
-import { RefSymbol } from '@vue/reactivity';
 import { defineStore, acceptHMRUpdate } from 'pinia';
 
 export interface LocoClient {
@@ -16,8 +15,6 @@ export const useLocoStore = defineStore('locos', () => {
     const definedLocos = ref(new Map<number, LocoClient>());
     const nameLookup = ref(new Map<string, number>());
     const activeThrottles: Ref<LocoClient[]> = ref([]);
-    // const activeThrottles: Ref<ActiveThrottle> = ref({});
-    console.log('Inside Store');
 
     const allLocos = computed(() => {
         return Array.from(definedLocos.value.values());
@@ -56,20 +53,8 @@ export const useLocoStore = defineStore('locos', () => {
         });
         const loco = definedLocos.value.get(locoId);
         if (loco) {
-            // activeThrottles.value[throttleId] = new Proxy(loco, handler);
             activeThrottles.value[throttleId] = loco;
         }
-    }
-    function setSpeed(throttleId: number, speed: number) {
-        activeThrottles.value[throttleId].speed = speed;
-        console.log(activeThrottles.value[throttleId]);
-        const loco = activeThrottles.value[throttleId];
-        // socket.emit(
-        //     'throttle/setSpeed',
-        //     loco.address.value,
-        //     loco.speed.value,
-        //     throttleId
-        // );
     }
     function addressFromName(name: string) {
         return nameLookup.value.get(name);
@@ -98,7 +83,6 @@ export const useLocoStore = defineStore('locos', () => {
         locoNames,
         addLoco,
         removeLoco,
-        setSpeed,
         switchActiveThrottle,
         lock,
         unlock,
@@ -108,20 +92,3 @@ export const useLocoStore = defineStore('locos', () => {
 if (import.meta.hot) {
     import.meta.hot.accept(acceptHMRUpdate(useLocoStore, import.meta.hot));
 }
-// const store = useLocoStore();
-
-// store.definedLocos.set(0, {
-//     name: 'Test',
-//     address: 0,
-//     speed: 0,
-//     direction: Direction.stopped,
-// });
-
-// store.activeThrottles[0] = store.definedLocos.get(0) || {
-//     name: 'Test',
-//     address: 0,
-//     speed: 0,
-//     direction: Direction.stopped,
-// };
-
-// store.activeThrottles[0].speed = 2;
